+++ src/components/PostListRow.js
@@ -4,6 +4,7 @@ import {
   TouchableHighlight,
   Image,
   Platform,
+  Navigator,
   Text,
   View,
 } from 'react-native';
